Race condition in session.DiskStore
Quite a few times I had the exception "EOFError: Ran out of input" being raised. It happened when I was sending multiple HTTP requests to the application.

I still do not understand why the application keeps re-writing the session files when I am not modifying any session value (not intentionally). But anyway, here are the messages presented when the problem happens:

  File "/opt/pf-connect/web/application.py", line 255, in process
    return p(lambda: process(processors))
  File "/opt/pf-connect/web/session.py", line 83, in _processor
    self._load()
  File "/opt/pf-connect/web/session.py", line 104, in _load
    d = self.store[self.session_id]
  File "/opt/pf-connect/web/session.py", line 251, in __getitem__
    return self.decode(pickled)
  File "/opt/pf-connect/web/session.py", line 210, in decode
    return pickle.loads(pickled)
EOFError: Ran out of input

After some investigation, I concluded that the session file was being written and read at the same time. As a result, the thread reading it was finding an incomplete file and raising the exception.

I solved the problem by adding a recursive mutex to the DiskStore class, and modifying the methods __setitem__ and __getitem__ so they acquire that mutex before trying to access the session files.